fix template instantiation for host function
