+++ libraries/chain/webassembly/context_free.cpp
@@ -1,4 +1,5 @@
 #include <eosio/chain/webassembly/interface.hpp>
+#include <eosio/chain/apply_context.hpp>
 
 namespace eosio { namespace chain { namespace webassembly {
    int32_t interface::get_context_free_data( uint32_t index, legacy_span<char> buffer) const {
