+++ libraries/chain/webassembly/crypto.cpp
@@ -1,6 +1,7 @@
 #include <eosio/chain/webassembly/interface.hpp>
 #include <eosio/chain/protocol_state_object.hpp>
 #include <eosio/chain/transaction_context.hpp>
+#include <eosio/chain/apply_context.hpp>
 
 namespace eosio { namespace chain { namespace webassembly {
