+++ libraries/chain/webassembly/privileged.cpp
@@ -3,6 +3,7 @@
 #include <eosio/chain/protocol_state_object.hpp>
 #include <eosio/chain/transaction_context.hpp>
 #include <eosio/chain/resource_limits.hpp>
+#include <eosio/chain/apply_context.hpp>
 
 #include <vector>
 #include <set>
